comit clases

+++ src/Spotify.Core/TipoSuscripcion.cs
@@ -3,7 +3,7 @@ namespace Spotify.Core;
 public class TipoSuscripcion
 {
     public int IdSuscripcion {get;set;}
-    public sbyte Duracion {get;set;}
-    public sbyte Costo {get;set;}
+    public uint Duracion {get;set;}
+    public uint Costo {get;set;}
     public required string Tipo {get;set;}
 }
